Add serde serialization support
Adds optional serializaiton support through serde.  Requires
the nightly version of serde (e.g. compiler plugins and macros)
instead of the syntex version.  The syntex version was too
invasive to easily add to Chrono without a lot of changes.

Because cargo cannot have a feature and a dependency share the same
name, the feature for this is called "serde_support" instead of
serde.