Remove or clarify references to time crate
The time crate will _hopefully_ have a v0.2 release coinciding with the
release of Rust 1.40. As such, all references to the time crate make
clear that it is the v0.1 crate, not the soon-to-be-released v0.2 crate.
The first line of the README was also changed, as it's not a strict
superset of the new version.