Making Duration creation functions const
This commit updates `Duration::seconds` and siblings to be const. This
requires the const-panic api that only stabilized in 1.57, so it may be
best to hide this behid a feature flag somehow so as to not require a
MSRV of 1.57. This is my first pr so I don't know the policy around
this.

This pr also adds a bit of mess to the functions because we can't use
`.expect("error")` or `Datetime::partial_cmp`. I think it is worth it
because const durations would be really nice to have.