Add niche optimization for `NaiveTime`
`NaiveTime` contains two integer fields, both with a restricted range:

- `secs`, the second of the day, is less than 86,400.
- `frac` is used for two things. To store the nanosecond of the second,
  and a leap second, so its highest valus is 1,999,999,999.

So, in order to allow niche optimization, one or both values could be
implemented in a type that does not allow higher values. Because `frac`
is stored later in the struct, it has the better option to yield
optimized code, e.g. when used with `Result<NaiveTime, E>`. The compiler
has a longer consecutive sequence of bytes to store non-NaiveTime its
data.