Implement generic assign operations for various types
Includes:
* `oldtime::Duration`
* `datetime::DateTime`
* `naive::date::NaiveDate`
* `naive::datetime::NaiveDateTime`
* `naive::time::NaiveTime`

Did not implement it for `date::Date`. This is because `Add` expects to be passed the owned object, but `AddAssign` only has access to a mutable reference. Since `Date` doesn't
implement `Copy`, that makes it impossible to generically and safely implement `AddAssign` using `Add` without cloning. It might be possible with unsafe code, but that's probably over
the top. `date::Date` is deprecated anyways.