Do not depend on `android-tzdata` if the `clock` feature is not enabled
All usages of `android-tzdata` types are within the `offset::local`
module, but such module is not built if the `clock` feature is not
enabled. Therefore, `android-tzdata` is effectively an unused dependency
when the `clock` feature is disabled.

To avoid downloading unnecessary crates on build time, let's make the
dependency on `android-tzdata` conditional on enabling the `clock`
feature, as it was done with e.g. `iana-time-zone`.